fixup! usbus: Initial work to a unified USB stack
